+++ khoros/objects/users.py
@@ -6,7 +6,7 @@
 :Example:           ``users.create(khoros_object, username='john_doe', email='john.doe@example.com')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     17 Jul 2020
+:Modified Date:     26 Dec 2020
 """
 
 import warnings
@@ -21,6 +21,9 @@ logger = log_utils.initialize_logging(__name__)
 def create(khoros_object, user_settings=None, login=None, email=None, password=None, first_name=None, last_name=None,
            biography=None, sso_id=None, web_page_url=None, cover_image=None):
     """This function creates a new user in the Khoros Community environment.
+
+    .. versionchanged:: 3.3.0
+       Updated ``khoros_object._settings`` to be ``khoros_object.core_settings``.
 
     .. versionchanged:: 2.7.4
        The HTTP headers were changed to be all lowercase in order to be standardized across the library.
@@ -53,7 +56,7 @@ def create(khoros_object, user_settings=None, login=None, email=None, password=N
     # TODO: Add functionality for followers, following, rank, roles, user_avatar and user_badges
     payload = structure_payload(user_settings, login, email, password, first_name, last_name, biography, sso_id,
                                 web_page_url, cover_image)
-    query_url = f"{khoros_object._settings['v2_base']}/users"
+    query_url = f"{khoros_object.core_settings['v2_base']}/users"
     headers = {'content-type': 'application/json'}
     response = api.post_request_with_retries(query_url, payload, auth_dict=khoros_object.auth, headers=headers)
     if not api.query_successful(response):
@@ -266,6 +269,9 @@ def structure_payload(user_settings=None, login=None, email=None, password=None,
 def delete(khoros_object, user_id, return_json=False):
     """This function deletes a user from the Khoros Community environment.
 
+    .. versionchanged:: 3.3.0
+       Updated ``khoros_object._settings`` to be ``khoros_object.core_settings``.
+
     :param khoros_object: The core :py:class:`khoros.Khoros` object
     :type khoros_object: class[khoros.Khoros]
     :param user_id: The User ID of the user to be deleted
@@ -276,7 +282,7 @@ def delete(khoros_object, user_id, return_json=False):
     """
     # TODO: Allow other identifiers (e.g. login, email, etc.) to be provided instead of just the User ID
     # TODO: Add a confirmation prompt that can be optionally disabled
-    query_url = f"{khoros_object._settings['v2_base']}/users/{user_id}"
+    query_url = f"{khoros_object.core_settings['v2_base']}/users/{user_id}"
     return api.delete(query_url, return_json, auth_dict=khoros_object.auth)